+++ cryptview/package.json
@@ -12,6 +12,7 @@
     "@testing-library/user-event": "^13.5.0",
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.1.0",
+    "history": "^5.3.0",
     "lucide-react": "^0.354.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
@@ -52,6 +53,7 @@
     "crypto-js": "^4.2.0",
     "lightweight-charts": "^4.1.3",
     "react-google-charts": "^4.0.1",
+    "react-router-dom": "^6.22.3",
     "tailwindcss": "^3.4.1"
   }
 }
