+++ cryptview/src/App.test.js
@@ -1,4 +1,4 @@
-import { render, screen } from '@testing-library/react';
+import { render } from '@testing-library/react';
 import App from './App';
 import { Provider } from 'react-redux';
 import { configureStore } from '@reduxjs/toolkit';
@@ -7,27 +7,25 @@ describe('Test Name', () => {
 	let renderComponent = () => null;
 	let store = null;
 
-	beforeEach(() => {
+	const reset = () => {
 		renderComponent = () => {
-			
+				
 			store = configureStore()
 			
-			render(
+			return render(
 				<Provider store={store}>
 					<App></App>
 				</Provider>
 			)
 		}
+	}
+
+	beforeEach(() => {
+		reset();
 	})
 
 	afterEach(() => {
 		store = null;
-	});
-
-	it('renders learn react link', () => {
-		render(<App />);
-		const linkElement = screen.getByText(/learn react/i);
-		expect(linkElement).toBeInTheDocument();
 	});
 
 });
