+++ cryptview/src/index.js
@@ -11,5 +11,5 @@ root.render(
 	  	<Provider store={store}>
     		<App />
 		</Provider>
-  </React.StrictMode>
+  	</React.StrictMode>
 );
